% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aggregate_coords_cpp}
\alias{aggregate_coords_cpp}
\title{Aggregate coordinates by component (mean pooling)}
\usage{
aggregate_coords_cpp(coords, component_labels, n_components)
}
\arguments{
\item{coords}{Numeric matrix (voxels x 3)}

\item{component_labels}{IntegerVector of component labels (0-based)}

\item{n_components}{Number of unique components}
}
\value{
NumericMatrix of aggregated coordinates (n_components x 3)
}
\description{
Computes mean coordinate for each component efficiently.
}
\keyword{internal}
