% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aggregate_features_cpp}
\alias{aggregate_features_cpp}
\title{Aggregate features by component (mean pooling)}
\usage{
aggregate_features_cpp(feature_mat, component_labels, n_components)
}
\arguments{
\item{feature_mat}{Numeric matrix (features x voxels)}

\item{component_labels}{IntegerVector of component labels (0-based)}

\item{n_components}{Number of unique components}
}
\value{
NumericMatrix of aggregated features (features x n_components)
}
\description{
Computes mean feature vector for each component efficiently.
}
\keyword{internal}
