% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acsc.R
\name{block_partition}
\alias{block_partition}
\title{Partition voxel coordinates into coarse blocks}
\usage{
block_partition(coords, block_size)
}
\description{
Partition voxel coordinates into coarse blocks
}
\keyword{internal}
