% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acsc.R
\name{build_acsc_graph}
\alias{build_acsc_graph}
\title{Build ACSC adjacency graph}
\usage{
build_acsc_graph(
  block_summary,
  ann_k,
  alpha,
  spatial_weighting,
  block_size,
  correlation_metric = c("pearson", "spearman", "robust"),
  knn_proj_dim = 0L,
  knn_proj_method = c("none", "dct", "rp"),
  knn_proj_seed = 1L
)
}
\description{
Build ACSC adjacency graph
}
\keyword{internal}
