% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rena_plus.R
\name{build_grid_adjacency}
\alias{build_grid_adjacency}
\title{Internal: build 3D grid adjacency for masked voxels (fast C++ version)}
\usage{
build_grid_adjacency(mask, mask_idx, connectivity)
}
\description{
Internal: build 3D grid adjacency for masked voxels (fast C++ version)
}
\keyword{internal}
