% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_clust.R
\name{cl_class_ids.cluster_result}
\alias{cl_class_ids.cluster_result}
\title{Extract Class IDs from Cluster Result}
\usage{
cl_class_ids.cluster_result(x)
}
\arguments{
\item{x}{A \code{cluster_result} object containing clustering information.}
}
\value{
An integer vector of cluster assignments, one for each data point.
}
\description{
This function extracts the cluster class identifiers from a cluster result object.
It is a method for the \code{\link[clue]{cl_class_ids}} generic from the \code{clue} package.
}
\seealso{
\code{\link[clue]{cl_class_ids}} for the generic function.
}
