% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster4d_init.R
\name{cluster4d_init}
\alias{cluster4d_init}
\alias{initialize_clusters}
\title{Initialization Methods for 4D Clustering}
\usage{
initialize_clusters(
  coords,
  features,
  n_clusters,
  method = c("gradient", "kmeans", "poisson", "grid", "random"),
  mask = NULL,
  vec = NULL,
  ...
)
}
\description{
Functions for initializing cluster seeds using various strategies.
}
\keyword{internal}
