% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_test.R
\name{compute_best_accuracy}
\alias{compute_best_accuracy}
\title{Compute Best Accuracy via Greedy Matching}
\usage{
compute_best_accuracy(cont)
}
\description{
Compute Best Accuracy via Greedy Matching
}
\keyword{internal}
