% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acsc.R
\name{compute_cluster_centroids}
\alias{compute_cluster_centroids}
\title{Compute centroids of each cluster}
\usage{
compute_cluster_centroids(labels, feature_mat)
}
\description{
Compute centroids of each cluster
}
\keyword{internal}
