% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acsc.R
\name{construct_block_label_array}
\alias{construct_block_label_array}
\title{Construct a 3D array of block labels}
\usage{
construct_block_label_array(block_id, mask)
}
\description{
Construct a 3D array of block labels
}
\keyword{internal}
