% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acsc.R
\name{cor_to_centroid}
\alias{cor_to_centroid}
\title{Correlate a voxel's time-series with a cluster centroid}
\usage{
cor_to_centroid(voxel_idx, lbl, feature_mat, cluster_centroids)
}
\description{
Correlate a voxel's time-series with a cluster centroid
}
\keyword{internal}
