% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acsc.R
\name{estimate_resolution}
\alias{estimate_resolution}
\title{Estimate Louvain resolution parameter}
\usage{
estimate_resolution(K, graph, seed = 20090601L)
}
\description{
Estimate Louvain resolution parameter
}
\keyword{internal}
