% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acsc.R
\name{expand_block_labels}
\alias{expand_block_labels}
\title{Expand block-level cluster labels to voxel level}
\usage{
expand_block_labels(cluster_result, block_id, mask.idx)
}
\description{
Expand block-level cluster labels to voxel level
}
\keyword{internal}
