% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbo_cluster.R
\name{filter_mat}
\alias{filter_mat}
\title{filter_mat}
\usage{
filter_mat(valmat, lp = 0, hp = 0.7, method = c("locfit", "bspline"))
}
\arguments{
\item{valmat}{a feature matrix with \code{nrow(valmat)} observations and `\code{ncol(valmat)} features.}

\item{lp}{the low-pass bandwidth parameter specified as a fraction
  of the number of time-points to include in each window. Must be between [0-1]
@param hp the high-pass bandwidth parameter specified as a fraction
  of the number of time-points to include in each window. Must be between [0-1]}
}
\description{
temporally filter a matrix of time-series
}
\examples{

valmat <- matrix(rnorm(200*10), 200, 10)
library(future)
plan(multicore)
svalmat <- filter_mat(valmat, lp=.2, hp=.7)
svalmat2 <- filter_mat(valmat, lp=.2, hp=.7, method="bspline")

}
