% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbo_cluster.R
\name{filter_vec}
\alias{filter_vec}
\title{filter_vec}
\usage{
filter_vec(bvec, mask, lp = 0, hp = 0.7, method = c("locfit", "bspline"))
}
\arguments{
\item{bvec}{an instance of type \code{\linkS4class{NeuroVec}} to filter}

\item{mask}{a binary mask defining the voxels to include in the filter operation}

\item{lp}{the low-pass bandwidth parameter specified as a fraction
  of the number of time-points to include in each window. Must be between [0-1]
@param hp the high-pass bandwidth parameter specified as a fraction
  of the number of time-points to include in each window. Must be between [0-1]}
}
\description{
bandpass filter a \code{\linkS4class{NeuroVec}} over the 4th (temporal) dimesion
}
