% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acsc.R
\name{find_boundary_voxels}
\alias{find_boundary_voxels}
\title{Identify boundary voxels}
\usage{
find_boundary_voxels(voxel_labels, nn_index)
}
\description{
boundary voxel = has at least one neighbor with a different label
}
\keyword{internal}
