% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_connected_components_cpp}
\alias{find_connected_components_cpp}
\title{Find connected components using Union-Find}
\usage{
find_connected_components_cpp(n_nodes, nearest_neighbor)
}
\arguments{
\item{n_nodes}{Number of nodes}

\item{nearest_neighbor}{IntegerVector of nearest neighbor for each node (0-based)}
}
\value{
IntegerVector of component labels (0-based, contiguous)
}
\description{
Finds weakly connected components in a directed graph defined by 1-NN edges.
}
\keyword{internal}
