% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acsc.R
\name{force_exact_k}
\alias{force_exact_k}
\title{Force clustering to have exactly K clusters by merging or splitting}
\usage{
force_exact_k(labels, feature_mat, K_target)
}
\description{
Force clustering to have exactly K clusters by merging or splitting
}
\keyword{internal}
