% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acsc.R
\name{index_to_grid}
\alias{index_to_grid}
\title{Convert linear indices to 3D grid coordinates}
\usage{
index_to_grid(mask, indices)
}
\description{
Convert linear indices to 3D grid coordinates
}
\keyword{internal}
