% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster4d_init.R
\name{init_random_seeds}
\alias{init_random_seeds}
\title{Random seed initialization}
\usage{
init_random_seeds(coords, n_clusters)
}
\arguments{
\item{coords}{Coordinate matrix}

\item{n_clusters}{Number of clusters}
}
\value{
List with seeds, initial_labels, and seed_coords
}
\description{
Randomly selects seed points.
}
\keyword{internal}
