% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervoxels.R
\name{knn_shrink}
\alias{knn_shrink}
\title{K-nearest-neighbor shrink}
\usage{
knn_shrink(bvec, mask, k = 5, connectivity = 27)
}
\arguments{
\item{bvec}{A \code{\link[neuroim2:NeuroVec-class]{NeuroVec}} instance (the data).}

\item{mask}{A \code{\link[neuroim2:NeuroVol-class]{NeuroVol}} mask defining the voxels to include. If numeric, nonzero = included.}

\item{k}{The number of nearest neighbors to average over.}

\item{connectivity}{The number of spatial neighbors to include in the search around each voxel.}
}
\value{
A \code{SparseNeuroVec} or similar object with the smoothed data.
}
\description{
Replace each voxel by the mean of its k nearest neighbors in its local spatial neighborhood.
}
\examples{
\dontrun{
mask <- neuroim2::NeuroVol(array(1, c(20,20,20)), neuroim2::NeuroSpace(c(20,20,20)))
bvec <- replicate(10,
                  neuroim2::NeuroVol(array(runif(20*20*20), c(20,20,20)),
                           neuroim2::NeuroSpace(c(20,20,20))),
                  simplify=FALSE)
bvec <- do.call(neuroim2::concat, bvec)
sbvec <- knn_shrink(bvec, mask, k=3)
}

}
