% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{make_dct_basis}
\alias{make_dct_basis}
\title{Generate DCT-II Basis Matrix}
\usage{
make_dct_basis(n_time, n_basis)
}
\arguments{
\item{n_time}{Number of timepoints}

\item{n_basis}{Number of DCT basis functions (including constant)}
}
\value{
Numeric matrix (n_time x n_basis) with orthonormal DCT basis
}
\description{
Creates a discrete cosine transform (type II) basis matrix for detrending.
}
\details{
DCT basis functions are excellent for removing low-frequency drift in fMRI.
The first basis is constant (mean), subsequent bases capture increasingly
higher frequency components.
}
