% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{make_poly_basis}
\alias{make_poly_basis}
\title{Generate Polynomial Basis Matrix}
\usage{
make_poly_basis(n_time, degree)
}
\arguments{
\item{n_time}{Number of timepoints}

\item{degree}{Polynomial degree (0 = constant, 1 = linear, 2 = quadratic, etc.)}
}
\value{
Numeric matrix (n_time x (degree+1)) with orthonormalized polynomial basis
}
\description{
Creates a polynomial basis matrix for detrending.
}
\details{
Uses Gram-Schmidt orthonormalization for numerical stability.
Degree 0 = mean removal, 1 = linear detrend, 2 = quadratic, etc.
}
