% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_clust.R
\name{merge_clus.cluster_result}
\alias{merge_clus.cluster_result}
\title{Merge Clustering Results for ClusteredNeuroVol Objects}
\usage{
\method{merge_clus}{cluster_result}(x, method = "SE", ...)
}
\arguments{
\item{x}{A \code{ClusteredNeuroVol} object or an object of class \code{"cluster_result"}.}

\item{method}{A character string indicating the consensus clustering algorithm to use. Default is "SE".
See \code{\link[clue]{cl_consensus}} for available methods.}

\item{...}{Additional clustering results to be merged.}
}
\value{
A \code{\linkS4class{ClusteredNeuroVol}} instance representing the consensus partition.
}
\description{
This method of merge_clus is specifically designed to merge clustering results for \code{ClusteredNeuroVol} objects
by performing consensus clustering across multiple clustering results.
}
\seealso{
\code{\link[clue]{cl_consensus}}, \code{\link[clue]{as.cl_hard_partition}}, \code{\link[clue]{cl_ensemble}}
}
