% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{meta_clust}
\alias{meta_clust}
\title{Meta Clustering for Cluster Results (S4 Generic)}
\usage{
meta_clust(x, cuts, ...)
}
\arguments{
\item{x}{A clustering result, typically an object of class \code{"cluster_result"}.}

\item{cuts}{The number of cluster cuts to consider.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Depends on the method called.
}
\description{
The meta_clust generic function is used to define methods for different
classes of input objects.
}
