% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acsc.R
\name{normalize_features}
\alias{normalize_features}
\title{Normalize feature matrix to unit-length vectors}
\usage{
normalize_features(feature_mat)
}
\description{
Normalize feature matrix to unit-length vectors
}
\keyword{internal}
