% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acsc.R
\name{preprocess_time_series}
\alias{preprocess_time_series}
\title{Preprocess fMRI time-series data}
\usage{
preprocess_time_series(bvec, mask, correlation_metric)
}
\description{
Preprocess fMRI time-series data
}
\keyword{internal}
