% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3s.R
\name{print.g3s_result}
\alias{print.g3s_result}
\title{Print Method for G3S Results}
\usage{
\method{print}{g3s_result}(x, ...)
}
\arguments{
\item{x}{A g3s_result object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for G3S Results
}
