% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{spatial_gradient}
\alias{spatial_gradient}
\title{Spatial Gradient Calculation}
\usage{
spatial_gradient(vol, mask, sigma = 0.5)
}
\arguments{
\item{vol}{A \code{NeuroVol} instance for which the spatial gradient should be calculated.}

\item{mask}{A \code{NeuroVol} mask defining the voxels to include in the spatial gradient calculation.
If the mask contains \code{numeric} data, nonzero values will define the included voxels.
If the mask is a \code{\link[neuroim2:LogicalNeuroVol-class]{LogicalNeuroVol}}, then \code{TRUE} will define the set
of included voxels.}

\item{sigma}{A numeric value controlling the spatial weighting function. Default is 0.5.}
}
\value{
A \code{NeuroVol} instance containing the spatial gradient values for the input \code{vol}.
}
\description{
The spatial_gradient function calculates the spatial gradient of a \code{NeuroVol} instance within the specified mask.
}
\examples{
\dontrun{
mask <- neuroim2::NeuroVol(array(1, c(20,20,20)), neuroim2::NeuroSpace(c(20,20,20)))
input_vol <- neuroim2::NeuroVol(array(runif(20*20*20), c(20,20,20)),
  neuroim2::NeuroSpace(c(20,20,20)))
gradient_vol <- spatial_gradient(input_vol, mask)
}

}
\seealso{
\code{\link[adjoin:spatial_laplacian]{spatial_laplacian}}, \code{\link[adjoin:weighted_spatial_adjacency]{weighted_spatial_adjacency}}
}
