% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervoxels.R
\name{supervoxel_cluster_time}
\alias{supervoxel_cluster_time}
\title{cluster objects with a temporal constraint}
\usage{
supervoxel_cluster_time(
  feature_mat,
  K = min(nrow(feature_mat), 100),
  sigma1 = 1,
  sigma2 = TR * 3,
  iterations = 50,
  TR = 2,
  filter = list(lp = 0, hp = 0),
  use_medoid = FALSE,
  nreps = 5
)
}
\arguments{
\item{K}{the number of clusters to find.}

\item{sigma1}{the bandwidth of the heat kernel for computing similarity of the data vectors.}

\item{sigma2}{the bandwidth of the heat kernel for computing similarity of the coordinate vectors.
If this value is small, then relatively larger weights are given to nearby voxels. If is is large, then
spatial weights will be less salient. A relatively large sigma1/sigma2 ratio weights data features more than
spatial features, whereas as large sigma2/sigma1 ration does the opposite.}

\item{iterations}{the maximum number of cluster iterations}

\item{use_medoid}{whether to use the medoids to define cluster centroids}
}
\description{
cluster objects with a temporal constraint
}
\examples{
feature_mat <- matrix(rnorm(100*10), 100, 10)
library(future)
plan(multicore)
cres <- supervoxel_cluster_time(t(feature_mat), K=20)
}
