% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbo_cluster.R
\name{turbo_cluster_time}
\alias{turbo_cluster_time}
\title{cluster objects with a temporal constraint}
\usage{
turbo_cluster_time(valmat, K = min(nrow(valmat), 100), sigma1 = 1,
  sigma2 = TR * 3, iterations = 50, TR = 2, filter = list(lp = 0, hp =
  0), use_medoid = FALSE, nreps = 5)
}
\description{
cluster objects with a temporal constraint
}
\examples{
valmat <- matrix(rnorm(100*10), 100, 10)
library(future)
plan(multicore)
cres <- turbo_cluster_time(t(valmat), K=20)

}
