% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbo_cluster.R
\name{compute_centroids}
\alias{compute_centroids}
\title{get centroids for a matrix and set of assignments}
\usage{
compute_centroids(valmat, grid, assignment, medoid = FALSE)
}
\description{
get centroids for a matrix and set of assignments
}
