% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rena_plus.R
\name{er_ml_rena}
\alias{er_ml_rena}
\title{Backwards-compatible alias for rena_plus}
\usage{
er_ml_rena(bvec, mask, ...)
}
\arguments{
\item{bvec}{A \code{NeuroVec} providing voxel time-series.}

\item{mask}{A \code{NeuroVol} mask; non-zero voxels are included.}

\item{...}{Reserved for future options.}
}
\value{
Same as \code{\link{rena_plus}}.
}
\description{
Backwards-compatible alias for rena_plus
}
