% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R, R/turbo_cluster.R
\name{merge_clus}
\alias{merge_clus}
\alias{merge_clus.cluster_result}
\title{merge a set of clustering results using a consensus clustering algorithm}
\usage{
merge_clus(x, ...)

\method{merge_clus}{cluster_result}(x, ...)
}
\value{
a \code{linkS4class{ClusteredNeuroVol}} instance
}
\description{
merge a set of clustering results using a consensus clustering algorithm
}
