% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_clust.R
\name{meta_clust.cluster_result}
\alias{meta_clust.cluster_result}
\title{Meta Clustering for Cluster Results}
\usage{
\method{meta_clust}{cluster_result}(x, cuts = min(as.integer(length(x$centers)/2), 2), ...)
}
\arguments{
\item{x}{A clustering result, typically an object of class \code{"cluster_result"}.}

\item{cuts}{The number of cluster cuts to consider. Default is the minimum
of half the number of centers and 2.}

\item{...}{Additional arguments:
\describe{
  \item{algo}{A character string indicating the clustering algorithm to use.
              Default is "hclust" (hierarchical clustering).}
  \item{hclust_method}{A character string specifying the agglomeration method
                       to use for hierarchical clustering. Default is "ward.D".}
}}
}
\value{
A list containing:
        \item{cvols}{A list of \code{\linkS4class{ClusteredNeuroVol}} instances.}
        \item{cuts}{The number of cluster cuts.}
        \item{cutmat}{A matrix representing the cluster assignments for each cut.}
        \item{hclus}{The hierarchical clustering result.}
}
\description{
The meta_clust function performs meta clustering on a given clustering result
by applying hierarchical clustering or other clustering algorithms.
}
\seealso{
\code{\link[stats]{hclust}}, \code{\link[stats]{cutree}}
}
