% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurocluster-package.R
\docType{package}
\name{neurocluster-package}
\alias{neurocluster}
\alias{neurocluster-package}
\title{neurocluster: Spatially Constrained Clustering for Neuroimaging Data}
\description{
Provides multiple algorithms for spatially constrained clustering of 3D and 4D neuroimaging data. Methods include supervoxels, SNIC (Simple Non-Iterative Clustering), SLIC (Simple Linear Iterative Clustering), slice-based minimum spanning forest, FLASH-3D with DCT compression, and adaptive correlation superclustering (ACSC). All methods balance feature similarity and spatial proximity to produce coherent brain parcellations. A unified 'cluster4d' interface standardizes access to all algorithms.
}
\author{
\strong{Maintainer}: Bradley Buchsbaum \email{brad.buchsbaum@gmail.com}

}
\keyword{internal}
