% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acsc.R
\name{refine_voxel_boundaries}
\alias{refine_voxel_boundaries}
\title{Refine voxel boundaries using cached cluster centroids}
\usage{
refine_voxel_boundaries(voxel_labels, feature_mat, coords, max_iter)
}
\description{
For each boundary voxel, compare correlation with each neighboring cluster's
cached centroid. This approach is much faster than comparing against all voxel time-series.
}
\keyword{internal}
